refactor: split app shell navigation
